+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor MainContainer to prevent unnecessary Airtable refetches by tracking last-fetched parameters
+- Refactor MainContainer to improve parameter comparison logic and prevent unnecessary Airtable refetches
 
 ---